# internal/openapidoc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "modelrouter API",
    "version": "0.1.0",
    "description": "OpenAI-compatible LLM routing proxy with admin APIs for configuration, health, limits, metrics, and usage logging."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/v1/models": {
      "get": {
        "tags": [
          "OpenAI Compatible"
        ],
        "summary": "List available models",
        "description": "Returns configured public model names. When client auth is enabled, the list is filtered by the caller's access group.",
        "operationId": "listModels",
        "security": [
          {
            "ClientBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Model list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/OpenAIUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/OpenAIError"
          }
        }
      }
    },
...
    "/admin/usage": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Query usage history",
        "description": "Requires metrics:read or admin:read. Reads local usage_log JSONL files and returns matching records with latest records first.",
        "operationId": "getAdminUsage",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientFilter"
          },
          {
            "$ref": "#/components/parameters/ModelFilter"
          },
          {
            "$ref": "#/components/parameters/RouteGroupFilter"
          },
          {
            "$ref": "#/components/parameters/EndpointFilter"
          },
          {
            "$ref": "#/components/parameters/FromUnix"
          },
          {
            "$ref": "#/components/parameters/ToUnix"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Usage history records.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRecordsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AdminUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/AdminForbidden"
          }
        }
      }
    },
    "/admin/usage/recent": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Query recent usage history",
        "description": "Requires metrics:read or admin:read. Returns recent usage records from local usage_log JSONL files.",
        "operationId": "getAdminUsageRecent",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientFilter"
          },
          {
            "$ref": "#/components/parameters/ModelFilter"
          },
          {
            "$ref": "#/components/parameters/RouteGroupFilter"
          },
          {
            "$ref": "#/components/parameters/EndpointFilter"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent usage records.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageRecordsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AdminUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/AdminForbidden"
          }
        }
      }
    },
    "/admin/usage/summary": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Aggregate usage history",
        "description": "Requires metrics:read or admin:read. Aggregates local usage_log JSONL records for charting.",
        "operationId": "getAdminUsageSummary",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientFilter"
          },
          {
            "$ref": "#/components/parameters/ModelFilter"
          },
          {
            "$ref": "#/components/parameters/RouteGroupFilter"
          },
          {
            "$ref": "#/components/parameters/EndpointFilter"
          },
          {
            "$ref": "#/components/parameters/FromUnix"
          },
          {
            "$ref": "#/components/parameters/ToUnix"
          },
          {
            "$ref": "#/components/parameters/Interval"
          },
          {
            "$ref": "#/components/parameters/Top"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated usage history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AdminUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/AdminForbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ClientBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Client API token from auth.keys[].key. Required when auth.enabled is true."
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin token from admin.token or admin.keys[].key."
      }
    },
    "parameters": {
      "ClientFilter": {
        "name": "client",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Filter by client name."
      },
      "ModelFilter": {
        "name": "model",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Filter by public model name."
      },
      "RouteGroupFilter": {
        "name": "route_group",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Filter by route group."
      },
      "EndpointFilter": {
        "name": "endpoint",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Filter by endpoint name."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 0,
          "maximum": 1000
        },
        "description": "Maximum returned items."
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0,
          "maximum": 1000000
        },
        "description": "Pagination offset."
      },
      "FromUnix": {
        "name": "from",
        "in": "query",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 0
        },
        "description": "Start Unix timestamp in seconds."
      },
      "ToUnix": {
        "name": "to",
        "in": "query",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 0
        },
        "description": "End Unix timestamp in seconds."
      },
      "Interval": {
        "name": "interval",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "minute",
            "hour",
            "day"
          ],
          "default": "hour"
        },
        "description": "Aggregation bucket interval."
      },
      "Top": {
        "name": "top",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        },
        "description": "Maximum number of breakdown items per dimension."
      }
    },
    "responses": {
      "OpenAIError": {
        "description": "OpenAI-style error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
      "OpenAIUnauthorized": {
        "description": "Invalid or missing client API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
...
      "AdminUnauthorized": {
        "description": "Invalid or missing admin token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminErrorResponse"
            }
          }
        }
      },
      "AdminForbidden": {
        "description": "Admin token lacks required permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ModelListResponse": {
        "type": "object",
        "required": [
          "object",
          "data"
        ],
        "properties": {
          "object": {
            "type": "string",
            "const": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelItem"
            }
          }
        }
      },
      "ModelItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "const": "model"
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "owned_by": {
            "type": "string",
            "example": "modelrouter"
          }
        }
      },
      "OpenAIErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "example": "invalid_request_error"
              },
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
      "AdminErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "UsageRecordsResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/UsagePageMeta"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageRecord"
            }
          }
        }
      },
      "UsageSummaryResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/UsagePageMeta"
          },
          "result": {
            "$ref": "#/components/schemas/UsageAggregateResult"
          }
        }
      },
      "UsageAggregateResult": {
        "type": "object",
        "properties": {
          "interval": {
            "type": "string",
            "enum": [
              "minute",
              "hour",
              "day"
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/UsageAggregate"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageBucket"
            }
          },
          "breakdown": {
            "$ref": "#/components/schemas/UsageBreakdown"
          }
        }
      },
      "UsageAggregate": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "integer",
            "format": "int64"
          },
          "successes": {
            "type": "integer",
            "format": "int64"
          },
          "failures": {
            "type": "integer",
            "format": "int64"
          },
          "error_rate": {
            "type": "number"
          },
          "bytes_out": {
            "type": "integer",
            "format": "int64"
          },
          "prompt_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "average_latency_ms": {
            "type": "number"
          },
          "average_end_to_end_token_rate": {
            "type": "number"
          },
          "average_generation_token_rate": {
            "type": "number"
          },
          "average_ttft_ms": {
            "type": "number"
          }
        }
      },
      "UsageBucket": {
        "type": "object",
        "properties": {
          "start_unix_sec": {
            "type": "integer",
            "format": "int64"
          },
          "end_unix_sec": {
            "type": "integer",
            "format": "int64"
          },
          "summary": {
            "$ref": "#/components/schemas/UsageAggregate"
          }
        }
      },
      "UsageBreakdown": {
        "type": "object",
        "properties": {
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageBreakdownItem"
            }
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageBreakdownItem"
            }
          },
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsageBreakdownItem"
            }
          }
        }
      },
      "UsageBreakdownItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/UsageAggregate"
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "returned": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "filters": {
            "type": "object",
            "properties": {
              "client": {
                "type": "string"
              },
              "model": {
                "type": "string"
              },
              "route_group": {
                "type": "string"
              },
              "endpoint": {
                "type": "string"
              }
            }
          }
        }
      },
      "UsagePageMeta": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PageMeta"
          },
          {
            "type": "object",
            "properties": {
              "filters": {
                "type": "object",
                "properties": {
                  "client": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "route_group": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "from": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "to": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        ]
      },
...
      "EventRecord": {
        "type": "object",
        "properties": {
          "unix_sec": {
            "type": "integer",
            "format": "int64"
          },
          "client": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "route_group": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "duration_ms": {
            "type": "number"
          },
          "bytes_out": {
            "type": "integer",
            "format": "int64"
          },
          "prompt_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int64"
          },
          "streaming": {
            "type": "boolean"
          },
          "ttft_ms": {
            "type": "number"
          },
          "end_to_end_token_rate": {
            "type": "number"
          },
          "generation_token_rate": {
            "type": "number"
          },
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "UsageRecord": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EventRecord"
          },
          {
            "type": "object",
            "properties": {
              "time": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
    }
  }
}